+++ packages/better_effect/lib/better_effect.dart
@@ -11,6 +11,7 @@ part 'src/di/resolver_backend.dart';
 part 'src/di/service_key.dart';
 part 'src/di/services.dart';
 part 'src/effect/effect.dart';
+part 'src/effect/effect_collection.dart';
 part 'src/effect/effect_context.dart';
 part 'src/effect/effect_local.dart';
 part 'src/effect/effect_ops.dart';